+++ common/core/src/main/java/is/codion/common/state/State.java
@@ -20,10 +20,8 @@ package is.codion.common.state;
 
 import is.codion.common.Conjunction;
 import is.codion.common.value.Value;
-import is.codion.common.value.ValueObserver;
 
 import java.util.Collection;
-import java.util.function.Consumer;
 
 /**
  * A class encapsulating a boolean state, non-nullable with null values translated to false.
@@ -104,11 +102,11 @@ public interface State extends StateObserver, Value<Boolean> {
 
 	/**
 	 * Creates a new {@link State} instance.
-	 * @param value the initial state value
+	 * @param initialValue the initial state value
 	 * @return a new {@link State} instance
 	 */
-	static State state(boolean value) {
-		return builder(value).build();
+	static State state(boolean initialValue) {
+		return builder(initialValue).build();
 	}
 
 	/**
@@ -119,11 +117,11 @@ public interface State extends StateObserver, Value<Boolean> {
 	}
 
 	/**
-	 * @param value the initial value
+	 * @param initialValue the initial value
 	 * @return a new {@link Builder} instance
 	 */
-	static Builder builder(boolean value) {
-		return new DefaultState.DefaultBuilder(value);
+	static Builder builder(boolean initialValue) {
+		return new DefaultState.DefaultBuilder(initialValue);
 	}
 
 	/**
@@ -207,60 +205,7 @@ public interface State extends StateObserver, Value<Boolean> {
 	/**
 	 * Builds a {@link State} instance.
 	 */
-	interface Builder {
-
-		/**
-		 * @param notify the notify policy for this value, default {@link Notify#WHEN_CHANGED}
-		 * @return this builder instance
-		 */
-		Builder notify(Notify notify);
-
-		/**
-		 * Adds a validator to the resulting state
-		 * @param validator the validator to add
-		 * @return this builder instance
-		 */
-		Builder validator(Validator<? super Boolean> validator);
-
-		/**
-		 * Links the given value to the resulting state
-		 * @param originalState the original state to link
-		 * @return this builder instance
-		 * @see Value#link(Value)
-		 */
-		Builder link(Value<Boolean> originalState);
-
-		/**
-		 * Links the given value observer to the resulting state
-		 * @param originalState the state to link
-		 * @return this builder instance
-		 * @see Value#link(ValueObserver)
-		 */
-		Builder link(ValueObserver<Boolean> originalState);
-
-		/**
-		 * @param listener a listener to add
-		 * @return this builder instance
-		 */
-		Builder listener(Runnable listener);
-
-		/**
-		 * @param consumer a consumer to add
-		 * @return this builder instance
-		 */
-		Builder consumer(Consumer<Boolean> consumer);
-
-		/**
-		 * @param weakListener a weak listener to add
-		 * @return this builder instance
-		 */
-		Builder weakListener(Runnable weakListener);
-
-		/**
-		 * @param weakConsumer a weak consumer to add
-		 * @return this builder instance
-		 */
-		Builder weakConsumer(Consumer<Boolean> weakConsumer);
+	interface Builder extends Value.Builder<Boolean, Builder> {
 
 		/**
 		 * @return a new {@link State} instance based on this builder